implement

+++ server/entities/entry.go
@@ -3,6 +3,6 @@ package entities
 // EntryResponse structure
 type EntryResponse struct {
 	ID    string `json:"id"`
-	Units int    `json:"units"`
+	Units int64  `json:"units"`
 	Start string `json:"start"`
 }
